+++ blurt_cpp_80211/interleave.cc
@@ -1,21 +1,21 @@
 #include "interleave.h"
 
-void interleave_permutation(int Ncbps, int Nbpsc, std::vector<int> &j) {
-    int s = Nbpsc/2;
+void interleave_permutation(size_t Ncbps, size_t Nbpsc, std::vector<size_t> &j) {
+    size_t s = Nbpsc/2;
     if (s < 1) s = 1;
     j.resize(Ncbps);
-    for (int k=0; k<Ncbps; k++) {
-        int i = (Ncbps/16) * (k % 16) + (k/16);
+    for (size_t k=0; k<Ncbps; k++) {
+        size_t i = (Ncbps/16) * (k % 16) + (k/16);
         j[k] = s * (i/s) + (i + Ncbps - (16 * i / Ncbps)) % s;
     }
 }
 
-void interleave_inverse_permutation(int Ncbps, int Nbpsc, std::vector<int> &k) {
-    int s = Nbpsc/2;
+void interleave_inverse_permutation(size_t Ncbps, size_t Nbpsc, std::vector<size_t> &k) {
+    size_t s = Nbpsc/2;
     if (s < 1) s = 1;
     k.resize(Ncbps);
-    for (int j=0; j<Ncbps; j++) {
-        int i = s * (j/s) + (j + (16*j/Ncbps)) % s;
+    for (size_t j=0; j<Ncbps; j++) {
+        size_t i = s * (j/s) + (j + (16*j/Ncbps)) % s;
         k[j] = 16 * i - (Ncbps - 1) * (16 * i / Ncbps);
     }
 }
