+++ blurt_cpp_80211/ofdm.h
@@ -5,9 +5,9 @@
 
 class OFDMFormat {
 public:
-    int nfft, ncp, ts_reps, Nsc, Nsc_used, preambleLength;
+    size_t nfft, ncp, ts_reps, Nsc, Nsc_used, preambleLength;
     std::vector<complex> sts_freq, sts_time, lts_freq, lts_time;
-    std::vector<int> dataSubcarriers, pilotSubcarriers;
+    std::vector<size_t> dataSubcarriers, pilotSubcarriers;
     std::vector<complex> pilotTemplate;
 };
