+++ blurt_cpp_80211/puncture.h
@@ -4,8 +4,8 @@ class PuncturingMask {
 private:
     std::vector<bool> mask;
 public:
-    int numerator, denominator;
-    PuncturingMask(int numerator, int denominator);
+    size_t numerator, denominator;
+    PuncturingMask(size_t numerator, size_t denominator);
     void puncture(const bitvector &input, bitvector &output) const;
     void depuncture(const std::vector<int> &input, std::vector<int> &output) const;
 };
