+++ blurt_cpp_80211/rate.h
@@ -6,9 +6,9 @@ class OFDMFormat;
 
 class Rate {
 public:
-    int encoding;
+    uint32_t encoding;
     Constellation constellation;
     PuncturingMask puncturingMask;
-    int Nbpsc, Nbps, Ncbps;
-    Rate(int encoding, const Constellation &constellation, const PuncturingMask &puncturingMask, const OFDMFormat &format);
+    size_t Nbpsc, Nbps, Ncbps;
+    Rate(uint32_t encoding, const Constellation &constellation, const PuncturingMask &puncturingMask, const OFDMFormat &format);
 };
