+++ blurt_cpp_80211/wave.h
@@ -3,6 +3,6 @@
 #include "blurt.h"
 
 bool readwave(const std::string &filename, std::vector<float> &output, float &Fs);
-bool writewave(const std::string &filename, const std::vector<float> &input, int Fs, int bytesPerSample, int nchannels);
+bool writewave(const std::string &filename, const std::vector<float> &input, size_t Fs, size_t bytesPerSample, size_t nchannels);
 
 #endif
